requirements.txt
flask
python-dotenv
tinytuya